Refactor code to improve accessibility and add dark mode toggle
The commit includes changes to improve accessibility by using the `useDir` composable instead of `useLDir` in `app.vue` and `index.vue`. It also adds a new component `TheDarkMode.vue` which provides a toggle button for switching between light and dark mode. The `tailwind.config.js` file was also updated to remove a path that was no longer needed.